delete unmanaged routed component
